Refactor: 폴더 이름 변경 PillChallenge > challenge
