+++ src/pages/Home/Challenge/PillManagePage.module.scss
