+++ src/pages/Home/Challenge/PillManagePage.tsx
