+++ src/pages/Home/Challenge/SupplementComponent/SupplementComponent.module.scss
