+++ src/pages/Home/Challenge/SupplementComponent/SupplementComponent.tsx
