+++ src/pages/Home/Challenge/components/InputClearButton.tsx
