+++ src/pages/Home/Challenge/components/IntakeDaySection.tsx
