+++ src/pages/Home/Challenge/components/IntakePeriodSection.tsx
