+++ src/pages/Home/Challenge/components/NameInputSection.tsx
